requirements.txt
langchain>=0.0.200
langchain-community>=0.0.10
python-dotenv>=1.0.0
langchain-openai>=0.0.2
eth-utils>=2.3.0
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0 

cryptography>=44.0.0
eth_utils
eth-hash[pycryptodome]
eth-account
web3>=6.0.0
siwe>=2.1.0

nillion-client
nada-dsl
betterproto==2.0.0b7